+++ packages/remixer/package.json
@@ -7,8 +7,8 @@
   "module": "dist/index.modern.js",
   "source": "src/index.js",
   "scripts": {
-    "build": "microbundle-crl --no-compress --format modern,cjs --external=react,react-dom,react-is",
-    "start": "microbundle-crl watch --no-compress --format modern,cjs --external=react,react-dom,react-is",
+    "build": "microbundle-crl --no-compress --format modern,cjs --external=react,react-dom,react-is --jsxFragment React.Fragment --jsx React.createElement",
+    "start": "microbundle-crl watch --no-compress --format modern,cjs --external=react,react-dom,react-is --jsxFragment React.Fragment --jsx React.createElement",
     "prepare": "run-s build",
     "test": "run-s test:unit test:lint test:build",
     "test:build": "run-s build",
