+++ packages/remixer/src/components/RemixTopbar.js
@@ -26,7 +26,9 @@ import Tooltip from '@mui/material/Tooltip';
 import UndoIcon from '@mui/icons-material/Undo';
 import { styled } from '@mui/material/styles';
 
-import { ClearDialog, ShareDialog, VisibilityDialog } from '.';
+import { ClearDialog } from './ClearDialog';
+import { ShareDialog } from './ShareDialog';
+import { VisibilityDialog } from './VisibilityDialog';
 import { RecursiveMenuItem, ShareIcon } from '@hyperaudio/common';
 
 const PREFIX = 'RemixTopbar';
